refactor: move tracing privacy validation out of the public observability API
